+++ src/Contracts/LocalizedRoute.php
@@ -9,8 +9,12 @@ interface LocalizedRoute
     /** @see \Goodcat\L10n\Mixin\LocalizedRoute::lang */
     public function lang(): Route;
 
+    /** @see \Goodcat\L10n\Mixin\LocalizedRoute::getKey */
+    public function getKey(): string;
+
     /**
      * @return Route[]
+     *
      * @see \Goodcat\L10n\Mixin\LocalizedRoute::makeTranslations
      */
     public function makeTranslations(): array;