+++ src/Events/PreferredLocaleUpdated.php
@@ -5,6 +5,7 @@ namespace Goodcat\L10n\Events;
 class PreferredLocaleUpdated
 {
     public function __construct(
-        public string $locale
+        public string $locale,
+        public ?string $previousLocale = null
     ) {}
 }