+++ src/L10n.php
@@ -3,15 +3,17 @@
 namespace Goodcat\L10n;
 
 use Goodcat\L10n\Contracts\LocalizedRoute;
-use Goodcat\L10n\Resolvers\BrowserPreferredLocale;
-use Goodcat\L10n\Resolvers\PreferredLocaleResolver;
-use Goodcat\L10n\Resolvers\UserPreferredLocale;
+use Goodcat\L10n\Contracts\LocalizedRouter;
+use Goodcat\L10n\Resolvers\BrowserLocale;
+use Goodcat\L10n\Resolvers\LocaleResolver;
+use Goodcat\L10n\Resolvers\SessionLocale;
+use Goodcat\L10n\Resolvers\UserLocale;
 use Illuminate\Routing\Route;
 use Illuminate\Routing\Router;
 
 class L10n
 {
-    /** @var PreferredLocaleResolver[] */
+    /** @var LocaleResolver[] */
     public static array $preferredLocaleResolvers;
 
     public function registerLocalizedRoutes(): void
@@ -24,24 +26,42 @@ class L10n
 
         foreach ($collection->getRoutes() as $route) {
             /** @var Route&LocalizedRoute $route */
+            if ($route->getAction('canonical')) {
+                continue;
+            }
+
+            $route->action['key'] = $route->getKey();
 
             foreach ($route->makeTranslations() as $localizedRoute) {
                 $collection->add($localizedRoute);
             }
-
-            $collection->add($route);
         }
     }
 
+    public function is(string ...$patterns): bool
+    {
+        /** @var LocalizedRouter&Router $router */
+        $router = app(Router::class);
+
+        $route = $router->current();
+
+        if ($canonical = $route?->getAction('canonical')) {
+            $route = $router->getByKey($canonical);
+        }
+
+        return $route && $route->named(...$patterns);
+    }
+
     /**
-     * @return PreferredLocaleResolver[]
+     * @return LocaleResolver[]
      */
     public static function getPreferredLocaleResolvers(): array
     {
-        if (!isset(self::$preferredLocaleResolvers)) {
+        if (! isset(self::$preferredLocaleResolvers)) {
             self::$preferredLocaleResolvers = [
-                new UserPreferredLocale,
-                new BrowserPreferredLocale,
+                new SessionLocale,
+                new UserLocale,
+                new BrowserLocale,
             ];
         }