+++ src/Resolvers/BrowserLocale.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Goodcat\L10n\Resolvers;
+
+use Illuminate\Http\Request;
+
+class BrowserLocale implements LocaleResolver
+{
+    public function resolve(Request $request): ?string
+    {
+        return $request->getPreferredLanguage();
+    }
+}