+++ tests/Feature/PreferredLocaleTest.php
@@ -2,40 +2,58 @@
 
 use Goodcat\L10n\L10n;
 use Goodcat\L10n\Middleware\SetPreferredLocale;
-use Goodcat\L10n\Resolvers\BrowserPreferredLocale;
-use Goodcat\L10n\Resolvers\UserPreferredLocale;
+use Goodcat\L10n\Resolvers\BrowserLocale;
+use Goodcat\L10n\Resolvers\SessionLocale;
+use Goodcat\L10n\Resolvers\UserLocale;
 use Goodcat\L10n\Tests\Support\User;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Session\Middleware\StartSession;
 use Illuminate\Support\Facades\Route;
 
 it('has default resolvers', function () {
     $resolvers = L10n::getPreferredLocaleResolvers();
 
     expect($resolvers)->toMatchArray([
-        new UserPreferredLocale,
-        new BrowserPreferredLocale,
+        new SessionLocale,
+        new UserLocale,
+        new BrowserLocale,
     ]);
 });
 
 it('detects preferred locale from browser', function () {
-    L10n::$preferredLocaleResolvers = [new BrowserPreferredLocale];
+    L10n::$preferredLocaleResolvers = [new BrowserLocale];
 
     Route::get('/example', fn () => 'Hello, World!')
-        ->lang(['it'])
         ->middleware(SetPreferredLocale::class);
 
-    $this->withHeader('Accept-Language', 'en')->get('/example');
+    $this->withHeader('Accept-Language', 'es')->get('/example');
 
-    expect(app()->getPreferredLocale())->toBe('en');
+    expect(app()->getPreferredLocale())->toBe('es');
 });
 
-it('detects preferred locale from user', function () {
-    L10n::$preferredLocaleResolvers = [new UserPreferredLocale];
+it('detects preferred locale from user', function (Authenticatable $user) {
+    L10n::$preferredLocaleResolvers = [new UserLocale];
 
-    Route::get('{lang}/example', fn () => 'Hello, World!')
-        ->lang(['en'])
+    Route::get('/example', fn () => 'Hello, World!')
         ->middleware(SetPreferredLocale::class);
 
-    $this->actingAs(new User)->get('en/example');
+    $this->actingAs($user)->get('/example');
 
-    expect(app()->getPreferredLocale())->toBe('en');
+    $expected = $user instanceof User ? 'en' : null;
+
+    expect(app()->getPreferredLocale())->toBe($expected);
+})->with([
+    'withPreferredLocale' => new User,
+    'withoutPreferredLocale' => new Authenticatable,
+]);
+
+it('detects preferred locale from the session', function () {
+    L10n::$preferredLocaleResolvers = [new SessionLocale];
+
+    Route::get('/example', fn () => 'Hello, World!')
+        ->middleware([StartSession::class, SetPreferredLocale::class]);
+
+    $this->withSession(['locale' => 'fr'])->get('/example');
+
+    expect(app()->getPreferredLocale())->toBe('fr');
 });