+++ tests/Support/lang/es/routes.php
@@ -2,4 +2,5 @@
 
 return [
     'example' => 'ejemplo',
+    'example.com' => 'es.example.com',
 ];